mec labs/requirements.txt
Flask==3.0.0
flask-pymongo==2.3.0
flask-bcrypt==1.0.1
pyjwt==2.8.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
scikit-learn==1.3.2
numpy==1.26.2
nltk==3.8.1
dnspython==2.4.2

